Plane: Use accurate roll limits if airspd. sensor
Aerodynamic load factor-based roll demand limits were only applied down
to ±25º to keep the aircraft maneuverable with a bad airspeed estimate.
However, with the addition of the AIRSPEED_STALL parameter, this is
unsafe if the aircraft is actually flying very close to its stall speed,
and can result in a spin.

This commit makes it possible for the roll demand limits to go all the
way down to 0º if an airspeed sensor is fitted and in use and the
AIRSPEED_STALL parameter is set, as under these conditions, the airspeed
estimate is accurate and there is very little bank angle overhead to
prevent a stall.