Plane: Fix pitch stick mixing full override
Fixed a bug where, in certain cases, the pitch setpoint couldn't be
fully overridden via stick mixing as intended. The issue occurred
because the pitch offset applied at full stick deflection was calculated
as twice the maximum pitch angle in the nudge direction, and since pitch
limits are often asymmetric, this could be insufficient.

For example, with the default pitch range of [-25, 20], if the commanded
pitch is -25º and the user nudges it fully upwards, the maximum
resulting pitch would be -25 + 20 * 2 = 15º, falling short of the
intended 20º.

Additionally, the non-linear stick input mapping has been replaced with
a linear one, which is more suitable for pitch control.