Plane: Add roll-only stick mixing
The current implementation of pitch stick mixing is unsafe, as it
introduces an attitude override that is decoupled from the TECS, and
also lacks any stall and overspeed prevention mechanisms. This means
that if a pitch nudge is held long enough, an overspeed or a stall will
occur.

This commit introduces a "FBW style (roll only)" setting to the
STICK_MIXING parameter. It behaves like the current "FBWMixing" option
for roll, but doesn't provide any pitch mixing. This is set as the
default for improved safety.

Pitch stick mixing can be re-enabled by setting STICK_MIXING to "FBW-A
style". This option replaces the removed "Direct" stick mixing mode in
the enum. A warning about the lack of flight envelope protections in
this mode has also been added to the description.